Move get/list/search into enginerequests.go with the other request methods
